lib: move Image class to lib/image.php
